Split Upload into provider/controller
